requirements.txt
tk
pydexcom
pystray
keyring
pillow
winshell
pywin32
pyinstaller